The linker binary should be build without the libc modifications needed for libhybris
As we're executing some android binaries on system boot to initialize the PowerVR graphics
chip and the linker binary is statically linked against the libc library it can't contain
the needed modifications for libhybris. Otherwise executing android binaries in a glibc
environment with

LD_LIBRARY_PATH=/system/lib /system/bin/ls

will fail with a SIGSEGV.

This path adjust the documentation for rebuilding everything accordingly.